Introduce Snowflake backend for Theia
Snowflake can be used as the data store for Theia, and to run
applications such as policy recommendation (using Snowflake virtual
warehouses). Users need to bring their own Snowflake and AWS accounts.

In this first commit, we add a new CLI tool (theia-sf), which can be
used to create the necessary cloud resources in Snowflake and AWS. The
CLI is built using Pulumi, an "infrastructure-as-code" tool, which is
itself built using Terraform providers. No third-party tool needs to be
installed directly by users, as theia-sf will install dependencies
itself in a temporary directory.

When running "theia-sf onboard", which will provision infrastructure,
users need to:
* configure AWS and Snowflake credentials
* provide an S3 bucket to store infrastructure state
* provide a KMS key to encrypt infrastructure state (notably secrets)

The advantage of using Pulumi to provision cloud resources is the
potential ability to support upgrades easily: resources can be created,
updated or deleted as needed whenever "theia-sf onboard" runs.

Pulumi is used to create a Snowflake database and schema. However, it is
not used to create objects (tables, views, etc.) within that schema. For
that we use a database migration tool. The rationale is that the Pulumi
/ Terraform provider for Snowflake doesn't support table schema updates
very well, and our current solution is more flexible. Once again, this
should enable us to support upgrades easily.

Signed-off-by: Antonin Bas <abas@vmware.com>