Add clusterUUID to flow table
To support multi-cluster, we add a new column "clusterUUID" as the
cluster identifier to flow table. This PR makes the corresponding
changes in the ClickHouse dataschema and related tests.

We found the ClickHouse init process is likely to fail due to timeout,
so we increase the env value of CLICKHOUSE_INIT_TIMEOUT from 12s to
60s, as suggested by Altinity documentation: https://kb.altinity.com/altinity-kb-kubernetes/altinity-kb-possible-issues-with-running-clickhouse-in-k8s/#clickhouse-init-process-failed

Signed-off-by: heanlan <hanlan@vmware.com>