Add API server cert controller
This change adds certificate controller to Theia manager. The public key
of API server TLS in case of self-signed, or CA cert in case of user
provided TLS, will be exposed to clients via configmap "theia-ca" in
flow-visibility namespace. This will allow cURL or client requests to be
made in "secure" fashion if the ca cert is added to trust chain.

The configmap will be updated when user provided TLS bundle is changed,
or the self-signed cert is rotated upon expiration.

Signed-off-by: Shawn Wang <wshaoquan@vmware.com>