Case insensitive string matching in Grafana plugins
When testing with snowflake datasource plugin, we found it will
automatically convert all fields name to upper case, which will
lead to string matching failure in our sankey-panel plugin and
chord-panel plugin, as we are doing case sensitive string matching.
As a work around, we do the following change to the panel plugins:
when search for fields with names in data frame, do case insenstive
string matching.

Signed-off-by: heanlan <hanlan@vmware.com>