Update Grafana Plugins
This PR does the following two things:

When testing with snowflake datasource plugin, we found it will
automatically convert all fields name to upper case, which will
lead to string matching failure in our sankey-panel plugin and
chord-panel plugin, as we are doing case sensitive string matching.
As a work around, we do the following change to both sankey and
chord plugins: when search for fields with names in data frame,
do case insenstive string matching.

When testing with snowflake datasource plugin, we found for a field
with type String, when the value is Null, it will not convert a
Null to ''. So we change our chord panel plugin to also check for
Null value, to be compatible with different datasources.

Signed-off-by: heanlan <hanlan@vmware.com>