Add Grafana dashboards with snowflake
This PR adds 4 Grafana dashboards with snowflake as the datasource:
Home Dashboard, Flow Records Dashboard, Pod-to-Pod Flows Dashboard,
and Network-Policy Flows Dashboard. These dashboards are basically
the same with ClickHouse dashboards in their contents, differ in
the query syntax. Currently, the snowflake dashboards are without
the ad-hoc filters, as the datasource plugin does not offer the
the support. We can revisit it later if we decide to work with
Grafana in the long-term.

In the deployment instruction, we suggest users deploy Grafana
with docker, so they do not need to install Grafana locally,
and it makes the custom plugins installation simple.

This PR should be merged after #112 and #118. The documentation
shall also be merged into the README.md introduced in #112

Signed-off-by: heanlan <hanlan@vmware.com>