Add RBAC auth and CA controller to Theia manager (#113)
This change adds k8s auth delegation to theia manager, and adds template
of cli service account / cluster role to allow access for specified API
groups and resources. A SA token secret is also added that can be used
for CLI to auth with Theia manager.

A certificate controller is also added to Theia manager. The public key
of API server TLS in case of self-signed, or CA cert in case of user
provided TLS, will be exposed to clients via configmap "theia-ca" in
flow-visibility namespace. This will allow cURL or client requests to be
made in "secure" fashion if the ca cert is added to trust chain.

The configmap will be updated when user provided TLS bundle is changed,
or the self-signed cert is rotated upon expiration.

Signed-off-by: Shawn Wang <wshaoquan@vmware.com>