Fix the ingestion error of snowflake on Pod label fields
In this commit, we fix the ingestion error of snowpipe on flow records
with Pod label fields. The reason behind this error is because commas
are used to separate different columns in CSV and Pod label columns
contain commas so they are separated into multiple columns when ingested
into snowflake.
To fix it, we add FIELD_OPTIONALLY_ENCLOSED_BY in COPY statement to use
single quote enclosing Pod label fields.

Signed-off-by: Yongming Ding <dyongming@vmware.com>