Add NetworkPolicy Recommendation on Snowflake Backend
In this commit, we add the NetworkPolicy Recommendation Application on
the Snowflake backend. NetworkPolicy Recommendation is implemented as
Snowflake UDFs and could be running on Snowflake warehouses. Users could
start it using command policy-recommendation in theia-sf CLI tool.

NetworkPolicy Recommendation UDFs are written in Python and stored under
udf/ directory.

Signed-off-by: Yongming Ding <dyongming@vmware.com>