Re-organize CHANGELOG for automatic updates
We introduce a similar workflow like Antrea which will take care of automatically updating the appropriate CHANGELOG file on the main branch whenever a new version of Theia (minor or bug fix) is released.
This will ensure that the CHANGELOG on the main branch is always up-to-date. The workflow cannot commit to the main branch directly (it is protected), but it will open a PR automatically.

Signed-off-by: Tushar Tathgur <tathgurt@tathgurtFLVDL.vmware.com>