Broadcast ns_allow_list in recommendation spark job (#418)
In this PR, we changed the allow namespace list to a PySpark Broadcast
variable. Previously, this allow namespace list is a global variable in
the Python file, but the MapReduce function couldn't get the latest
value if it's changed by user defined allow namespaces. So we choose to
use Broadcast variable to share the latest allow namespace list to the
distributed running map functions.

Signed-off-by: Yongming Ding <dyongming@vmware.com>