Pre-load images when creating cluster for multi-cluster e2e test
In this PR, we do:

Add --image flag when creating cluster for multi-cluster e2e test.
Remove Theia Manager when generating manifest for multi-cluster e2e test.
The reasons are:

During the creation of the multi-cluster end-to-end test, images were not preloaded, which caused an error. This error stated that the ClickHouse Pod's status couldn't become ready within the expected timeframe due to the extended duration required for pulling the images.
Theia Manager is not required in multi-cluster e2e test.

Signed-off-by: Yun-Tang Hsu <hsuy@vmware.com>